Bluetooth: Fix L2CAP control bit field corruption
When resending an I-frame, ERTM was reusing the control bits from the last
time it was sent, that was causing a corruption in the new control field
due to it dirty fields.

This patches extracts only the SAR bits from the old field and reuse it to
resend the packet, the others bits should be reset and receive the
updated value.

Signed-off-by: Gustavo F. Padovan <padovan@profusion.mobi>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit 95ffa97827371ede501615d9bd048eb5b49e8fe1)

Change-Id: Ibfba27e7dd4e73d48e7eebb4c73f9c8f9b338976
Signed-off-by: Mat Martineau <mathewm@codeaurora.org>