ARM: 6272/1: Convert L2x0 to use the IO relaxed operations
This patch is in preparation for a subsequent patch which adds barriers
to the I/O accessors. Since the mandatory barriers may do an L2 cache
sync, this patch avoids a recursive call into l2x0_cache_sync() via the
write*() accessors and wmb() and a call into l2x0_cache_sync() with the
l2x0_lock held.

Change-Id: Iafd652eb7bce5ec97be25b6a1b0c8edf02602054
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
[pratikp@codeaurora.org: skip l2x0_init changes due to code divergence]
Signed-off-by: Pratik Patel <pratikp@codeaurora.org>