Bluetooth: Simplify L2CAP Streaming mode sending
As we don't have any error control on the Streaming mode, i.e., we don't
need to keep a copy of the skb for later resending we don't need to
call skb_clone() on it.
Then we can go one further here, and dequeue the skb before sending it,
that also means we don't need to look to sk->sk_send_head anymore.

The patch saves memory and time when sending Streaming mode data, so
it is good to mainline.

Signed-off-by: Gustavo F. Padovan <padovan@profusion.mobi>
(cherry picked from commit ccbb84af28594e19fd4bf27ff2828c80d03b6081)

Change-Id: If4644a561b812b6d37a0a1925bf192a96f8620be
Signed-off-by: Mat Martineau <mathewm@codeaurora.org>