ASoC: 8x60: Add device switch support for PCM sessions.
The PCM session is to be used for playback/capture and
needs to be routed to different sinks/sources. Add support to
switch dynamically between different sinks/sources.

Remove current assignment of the index to simple controls as they
are not relevant and should be used by the device controls to
identify the device number.

Add support for multisession for playback.

Change-Id: Ieb220620c7dbacf9f2335d6f037273fc5f04adf3
Signed-off-by: Asish Bhattacharya <asishb@codeaurora.org>