msm: clock-pcom: Ignore errors on max rate requests for EBI1
Requesting a rate of UINT_MAX will max out the ebi1 frequency.
Due to a bug on the remote side, the rate is actually maxed out
but an error is returned nonetheless. Since we can't fix the
firmware, just ignore errors when we know we're going to cause
buggy error reporting on the remote side.

CRs-Fixed: 272028
Change-Id: I6409a5750d24b03757725bfd35e6d9efa215745b
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>