msm: dma: support dma power management on msm8x60
Move the initialization currently done in the dma driver's init function
to a probe function, and make the datamover a proper platform_device on
all targets. This is done so that each adm has its own struct device
that it can pass to the clock api. Additionally, move the irq and base
address fields of the dmov_conf structure to platform data, so that
there is no need to rely on consistent irq number and base address
macro naming.

This change makes it so we actually vote for the ADM clocks on msm8x60,
rather than relying on them coming up on.

Change-Id: Ic5aaea3160e60848979f81e74bcc2f129cda0af7
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>