pmic8058-charger: wakeup on usb insertion
When the system goes to suspend state, USB phy is put in low power
mode. It means USB phy wont be able to detect interrupts.

The pmic charger driver is expected to wake the system up when usb is
inserted.  Make the usb interrupt in the charger as wakeup interrupt.

Change-Id: I3f51c25ae9adffaa36ed53f9bc82c790fca9efc4
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>