Revert "msm: acpuclock-8x60: Vote for PLL8 to be enabled when in use"
This reverts commit cfc475b4c992c531a51e95b42781668d39d543b5.

The pm-8x60 driver doesn't switch the CPUs to their AFAB sources
until after the decision has been made to allow/disallow XO
shutdown.  Because of this, votes for PLL8 in the acpuclock driver
may prevent XO shutdown. Since PLL8 should be on anyway, whether
we vote for it or not, don't bother voting for it at all.

Change-Id: Iafe652729d51b7e3305eb241ca614c0a3e0131b7
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>