msm: 8x60: Remove "always_on" vote for 8901_L5 regulator
8901_L5 regulator powers eMMC card's VCC voltage. During system
suspend it is not safe to disable this regulator as the card may be
in unknown state during wakeup and card detection failure happens.
SDCC driver is now taking care of this by not voting to disable this
regulator but keep in low power mode with load current 9mA. Hence, remove
always on vote for 8901_L5 regulator and let the driver decide the mode of
operation.

Change-Id: Ie96ef6d4e563f9f4d643ed3cba4f28223670c75b
Signed-off-by: Sujith Reddy Thumma <sthumma@codeaurora.org>