msm: rpm-regulator: Add new pin function options
The RPM code has added two new pin function enum values and
modified the request requirements when not using pin control.

Add the two new pin function options and make the default
pin function request value NONE instead of ON/OFF when pin
control is not in use.

Also, make sure that pin control and pin function values are
set to NONE when a regulator is disabled and reset when it
is re-enabled.

Change-Id: Ie3e66e850cf65699930f69b707a8a4b05d29cbeb
Signed-off-by: David Collins <collinsd@codeaurora.org>