spi_qsd: Use FIFO mode when cs_change is set and bpw is not 8, 16 or 32
When cs_change is specified and bits per word is not 8, 16 or 32
then we use FIFO mode. DM mode cannot be used for this case
as the ADM transfers in blocks of 16 bytes which the SPI core
is unable to understand. This fix helps some devices like the
camera sensor, to achieve optimal performance.

Change-Id: I44ad8a3839cabd4cb26494c3dc2514537505e88d
Signed-off-by: Harini Jayaraman <harinij@codeaurora.org>