usb: host: pehci: Set appropriate URB status in dequeue function
As part of unlinking of the URBs, PEHCI driver should set the
status of the URB as ENOENT (passed from hcd) in dequeue
function. But, PEHCI driver currently sets the URB status as
SUCCESS which results in client drivers wrongly assuming the
transfer as successful.
Also, USB core sets urb->reject during unlinking process or
when the URB is poisoned. Hence, return error in enqueue
function if resubmission of a URB is tried during unlinking.

Change-Id: I369db5c547e044015ccf4eb55f57b1775f0b1180
Signed-off-by: Manu Gautam <mgautam@codeaurora.org>