mfd: pm8xxx: Add irq support
Add support for the irq controller in Qualcomm 8xxx pmic. The 8xxx
interrupt controller provides control for gpio and mpp configured as
interrupts in addition to other subdevice interrupts. The interrupt
controller also provides a way to read the real time status of an
interrupt. This real time status is the only way one can get the
input values of gpio and mpp lines.

Change-Id: Ia69b669dffa90fb9588d31e4d17e6eda3cc00708
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>