msm: camera: switch power supply for FLUID 12M camera and D02.
LVSW1 doesn't have enough driving strength for the Sony 12M sensor,
its voltage drops below the 1.7V required by the sensor. The solution
is to modify the camera daughter board so that GP16 turns on as
digital voltage for main camera and D02, and GP6 turns on secondary
camera.

Change-Id: I09578b48939b6fde0972b41f28b4018c432aea4e
CRs-Fixed: 253514
Signed-off-by: Jilai Wang <jilaiw@codeaurora.org>