msm: clock-8x60: Initialize voting registers for HW-voteable clocks
The clock driver doesn't use SC1's voting register to control
HW-voteable clocks.  Zero it out so that disabling bits in the
SC0 register will cause the corresponding clocks to be disabled.

sc_aclk, sc_hclk, and the ADM clocks still need to be left on
at all times, until other subsystems begin properly voting for
them. The rest of the clocks should now correctly disable when
not needed.

Change-Id: If9f555604dd51b1571d3475d7dcabf13da612e79
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>