arm: 1M align init, text, rodata if CONFIG_STRICT_MEMORY_RWX is set
Init code, text, rodata and data need different permissions
and so they need to be on different pages. The kernel 1-to-1
mapping is constructed using 1M pages to improve TLB performance
and this should not be changed (to 4K pages). Therefore
ensure that each of these regions starts on a 1M boundary.

Change-Id: I4f6c990da16ff3b5bf1127ddd92c82172f15184b
Signed-off-by: Larry Bassel <lbassel@codeaurora.org>