USB: msm_otg: Fix the bug in PHY suspend failure recovery
When PHY suspend fails, OTG state machine is initialized to UNDEFINED
state and work is scheduled.  State machine work takes care of resetting
the PHY and initializing the inputs.  Peripheral or host controller
driver is activated based on the B-session valid and Id inputs.  But this
recovery mechanism with wall charger stops charging and activates the
peripheral controller driver which was already in active state.  Handle
the wall charger scenario in UNDEFINED state by changing the state to
B_PERIPHERAL and schedule the work to take care of resetting the PHY and
putting the hardware in low power mode.

CRs-Fixed: 274343
Change-Id: I88d2d284e78664291ac11ea724c8a9de42c2c796
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>