power: pmic8058: wait a second for chg completion
The CHG_DONE interrupt doesnt necessarily mean that the pmic has
finished charging. This can happen when the phone is idle consuming
current just above the termination limit. If one plugs the external charger
now the current jumps down and the CHG_DONE interrupt is triggered.

Wait for a second before taking the steps for chg completion. If an
external charger was plugged in the pmic will be requested to stop
charging within a second.

CRs-fixed: 274345
Change-Id: I0c133d1e4e6893d51382bb11e39a15d229b420a2
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>