power: pmic8058-charger: use non interruptible sleep
If the current thread is signalled, wait_for_completion_interruptible()
could return -ERESTARTSYS resulting in the batt_read_adc function
returning without executing the rest of the adc apis. Note that the
completion structure passed to the request_conv() function is on the
stack.

This causes issues later when the adc code gets to notifying this
request and ends up using a completion structure long gone.

CRs-fixed: 274344
Change-Id: Ie665c2a1c43541d9552e642e54c0a7c864928327
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>