crypto: Fix data structure used for sha operation
QCRYPTO and QCEDEV share the same QCE layer to communincate with the
CE hardware.  The data strcuture returned by QCE layer to the call back
to qCEDEV was not ahash_request data structure but qce_sha_req.
Although this works for the SHA imjplementation in QCEDEV layer,
it will not work when SHA support is implemented in QCRYPTO layer.

QCRYPTO uses ahash_data structure to store the relevant data needed to
communicate with the kernel crypto layer. ahash_data structure has entries
to scatter list info used in storing the data needed to be hashed and other
information related to how the sha_ctx information is stored (not available
from qce_sha_req)

In order to have the common QCE layer cater to both QCEDEV and
QCRYPTO, with regards to sha support, we should use a common ahash
data structure (as was done for the cipher implementation). ahash_request
dat structure should be returned to the caller.

Change-Id: Icd47f5895ce4483cc8a3cdefb15e3b66ba68dd61
Signed-off-by: Mona Hossain <mhossain@codeaurora.org>