msm: kgsl: sync outer cache when submitting ringbuffer commands
On msm 7x27, the outer cache must be synced as part of the mb()
macro. Otherwise the ringbuffer can be in an inconsistent state
in RAM when the write pointer is updated.

This is an interim change until we can safely disable
CONFIG_ARCH_HAS_BARRIERS and use the arm version mb() which calls
outer_sync() instead of the msm version which does not.

Change-Id: If65ff33c3450df250f56e3bdfd00cde34237271e
CRs-fixed: 273669
CRs-fixed: 274294
Signed-off-by: Jeremy Gebben <jgebben@codeaurora.org>