msm: clock-8x60: Take software control of IJPEG_AXI_CLK, and IMEM_AXI_CLK
These clocks support dynamic hardware gating but require core-
specific configuration to enable it.  Without this configuration,
the clocks are stuck on.  Since the configuration needs to be
done by device drivers that may or may not be present, use
software control for these clocks.  This allows the clocks to
remain off if the device drivers never load.

Change-Id: I7fda048999622e4792cd32cbd025ca0d1ded4042
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>