USB: rmnet-sdio: Fix bug in handling dmux_write_done flag
dmux_write_done flag is cleared before a write request is submitted
to sdio_dmux layer.  This flag is set after the write request is
complete.  But if the cable is disconnected while a write request is
pending dmux_write_done is not set after write request completion.
Hence set dmux_write_done before checking the cable connect status.
Otherwise write requests are not handled when the cable is connected
next time.

CRs-Fixed: 276074
Change-Id: I7ea49d736cb2573ed8716fe6e3ce23128ca8f027
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>