usb: gserial_sdio: Check sdio pipe availability before queuing pkts
sdio al driver returns error if channel is full and client drivers
try to write to the channel. Add checks in the driver to check for
channel availability before writing to it. Also do not remove the
request from write queue in case of any error while writing to sdio
channel.

Change-Id: I148021d57955ef3163dc1287b266a0ce06a40d0d
Signed-off-by: Vamsi Krishna <vskrishn@codeaurora.org>