msm: clock-local: Enable/disable power rails correctly
The power rails can be disabled early now that the reference
count has been pushed down into clock-local.c. This can lead to
power rails being turned off when they shouldn't be. This would
have been a problem before the pushdown if the clocks with power
rails weren't leaves in the tree.

Change-Id: I0930976c37026f1e80cd6aee07c257a5bddc93a6
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>