msm: clock-local: Select low-power sources when clocks are disabled
Power savings can be achieved by selecting un-clocked or slowly-
clocked inputs of the MUXs which feed clock generation hardware,
even when the branch and root clocks are already disabled.

When disabling a clock, also switch its MUX to a low-power
source. Similarly, for dual-banked clocks, select a low-power
source for the inactive bank when the clock is running.

CRs-Fixed: 274420
Change-Id: I000f94d7103558d2650482d2bf4fd3903deb2362
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>