msm: saw-regulator: Assume that PMIC 8901 is v2 when version unavailable
v1 PMIC 8901 chips require setpoint scaling when setting SMPS voltage
in order to compensate for hardware issues.  This scaling is not
required on v2 (and greater) PMIC 8901 chips.  The saw-regulator
handles this scaling for the scorpion core rails.  However, it is
limited during boot to ignoring voltage requests above the assumed
boot loader value until the PMIC 8901 version can be checked.  No
commercial devices should be using v1 PMIC 8901 chips.  Therefore,
change the saw-regulator driver to make the assumption that the PMIC
chip is v2 when the version can not yet be checked.

Change-Id: Iaa2efd03d8419e9be4f7b0852acd38a6612112a8
Signed-off-by: David Collins <collinsd@codeaurora.org>