power: pmic8058-charger: force VCHG to 0 volts only once
While going to suspend, when the usb cable is removed, multiple
invocations of chgval interrupt were seen. This chgval interrupt is a
wakeup interrupt which causes the system to wakeup from suspend.

In the chgval interrupt we write to test register to force VCHG
to 0 volts. This forcing further causes activation of chgval interrupt.
So force the VCHG to 0 volts only the first time we detect that the
cable is removed. Ignore the rest of the invocations.

CRs-Fixed: 274343
Change-Id: Ifad3054e7b113d1be689c0e94128299b2250f1bd
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>