ASoC: msm8x60: Fix the afe_loopback triggering issue
The driver uses two variables to store the current dev_ids configured
for afe loopback. However these variables are not explicitly set to a
value and thus they result in initialized to zero, which is a valid
device id for handset_rx.
Due to this when we enable/disable the handset_rx, it is triggering
AFE_Enable. This is fixed by initializing these variables to 0xFFFF
to ensure they do not match with any valid device ids.

Change-Id: I8bdc7f5ac6d9bd339c2e675a83f6e829aebb39d1
Signed-off-by: Deepa Madiregama <dmadireg@codeaurora.org>