msm: sdio: move sdio_al_setup to the device probe procedure
Due to historical reasons, when there was no detection mechanism
between the modem and the msm SDIO components, sdio_al had to
wait before calling sdio_al_setup, in order to allow the modem to
finish its initialization process.
Now, this is not required and it's more accurate to call sdio_al_setup
when the device is detected and initialized.

Change-Id: I3ddacf0cada0e635144e6f1fa05c1b947b1a9194
Signed-off-by: Maya Erez <merez@codeaurora.org>