rtc: rtc-msm: Add functionality to maintain sleep time.
In case when APPS gets time update from modem while APPS is in suspend
state, it looses the sleep time and thus alarm data structure is not
updated with the time spent in the suspend state.

Change in rtc-msm and rpc_server_time_remote driver to mantain sleep
time and update alarm structure accordingly.

CRs-Fixed: 271409
Change-Id: Icabea2f477afcd9e4e88e82025a3aeea6e532e3b
Signed-off-by: Ashay Jaiswal <ashayj@codeaurora.org>