dcc_tty: Optimize SMP fix
The DCC console driver was modified to work on SMP
platforms. Revert to the original version and to
provide SMP support in a lighter, more optimized
way, use a combination of hrtimers and workqueues
instead of workqueues alone.

The feature of hrtimers that makes this possible
is that a hrtimer scheduled on one core is guaranteed
to run on that core.

Change-Id: I0d3f5a098d7f44242bbd07c7209d3c43729f8bdd
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>