GIC: implement disable callback
The disable callback is required to be implemented so that we indicate
to the mpm that a particular interrupt has been disabled. Doing this
helps the mpm stay current with interrupt status allowing it to choose
the right low power mode while going in to idle.

Note that we do not mask the interrupt in hardware, taking advantage of
lazy disabling and IRQ_PENDING checks in check_wakeup_irqs() while
going to suspend.

Change-Id: Iae7bc1fe07c214ef6fc73542660297f0acda305e
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>