radio: tavarua: Fix enum warnings
The xfr_status variable is being read from an 8-bit unsigned register.
Although the value definitions it is compared with are defined in an
enum, these use bit arithmetic to define the cases (setting the high
bit) which brings the constant values outside of the range of the
enum.  This causes warnings on gcc 4.5.

Fix this by keeping the value as an unsigned char.

Change-Id: Iab1ebe7692f62ba5a1716aca9dcf9690c65dcc5f
Signed-off-by: David Brown <davidb@codeaurora.org>