msm: watchdog: Add parameter to configure who handles bark interrupts
Add an "appsbark" flag that determines who handles watchdog barks that
come in. If it is set to 0(the default), the secure side will handle
barks, and save the registers out to a buffer in the ramdump. If it is
set to 1, Linux will handle barks, and dump the stack traces of all
processes upon receiving a bark.

Additionally, the secure side requires that the interrupt be configured
before sending the scm call.

Change-Id: I23e43cdfab3474eaeab153d93574bd4668408304
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>