USB: f_mass_storage: Disable write cache support
As part of mode sense CBW command, communicate to the
host that write cache support is enabled and due to this
during the write commnd, the host is asking for FUA and
because of this write performance is degrading. Hence during
mode sense command, intimate the host that write cache is not
supported.

Change-Id: I62b352faf034c582b2472184c826b030a3a057a0
CRs-Fixed: 278310
Signed-off-by: Chiranjeevi Velempati <cvelempa@codeaurora.org>