usb: msm_otg: Fix BSV/ID checks if OTG_MODE is controlled by USER
OTG driver depends on ID and BSV changes in OTGSC register for
switching to host or peripheral modes. But, on some targets
OTG driver depends on USER input via debugfs to switch between
the modes. For such cases, handle the ID/BSV checks in the driver
based on the state-machine input flags.

Change-Id: I3ba79a69975f3cf43d64f4a206915742ffbb53da
Signed-off-by: Manu Gautam <mgautam@codeaurora.org>