msm: rpc_hsusb: Remove limitation on serial number string length
The current code assumes that serial number is always 17 bytes
(including null character).  Modem processor expects the serial
number string to be 32 bit aligned. So it is stored in a statically
allocated array of 20 byte size.  The RPC call was failing when
serial number string of size not equal to 17 is passed.  This patch
removes the limitation on serial number string length by dynamically
allocating the memory.

CRs-Fixed: 277996
Change-Id: I2177951268bd08d6a67e2b158ecfa0fc9246faab
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>