mtd: msm_nand: Fix wrong offset of erased page marker in 7x27A
In newer version of NAND controller operating in legacy 4bit RS ECC
when an erased page is detected the ECC engine tries to correct the
bytes and writes 0x54 to a particular offset. In older targets this
offset is at 0x3 for each codeword but in newer version this offset
is changed to 0xAF. Fix this by checking 0x54 for new offset also.

Change-Id: I5936f4ff9a90808a57134e69a22ac68627ce980f
Signed-off-by: Sujith Reddy Thumma <sthumma@codeaurora.org>