msm: acpuclock-8x60: Check redundant speed bin field
If a part has been reclassified as a member of a different speed
bin, the primary speed bin field will be changed to 0xF. This
indicates that software should use the separate 'redundant'
speed bin field instead of the primary field. If both fields
are 0xF, then fall back to the most conservative speed bin.

Change-Id: I18410a5b07d28754275b0b971f8130b806372c23
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>