msm: charm: Send graceful notification on reset
During reboot, it is necessary to set the status GPIO low
so the charm modem can prepare appropriately. This involves waiting
several seconds for the charm to respond by setting its own gpio
low. This wait must be blocking since this is a shutdown path
(otherwise we could shutdown without setting the GPIO), so
the watchdog must be pet while waiting for the charm to respond.

Change-Id: Icd9cfe3a22f54fae03eaaaaa806edd671e3fce7b
CRs-Fixed: 277711
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>