msm: clock-7x30: Select LPXO source when disabling clocks with pwr rails
For clocks that have modem-controlled power rails associated with them,
a clocked input must remain selected even when the clock is disabled.

The power rail APIs are meant to be called while the clock is disabled.
Because the modem may assume that a valid clock source is selected when
it performs the power rail control operations, the clock driver should
not select a grounded input. Instead, when disabling the clock, LPXO
will be selected for good power savings.

CRs-Fixed: 280184
Change-Id: If09c7e85fb65d6626e10f5d86592db2876e53755
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>