usb: rmnet_sdio: Set zero len flag for control ep in transfers
RmNet protocol requires zero length packet if data on control
endpoint is smaller than transfer length and is in multiples of
endpoint's max pkt size and. Set the zero length flag appropriately
and register for completion handler on control control IN transfers.
As the host is requesting for max pkt size of 4096, increase max
setup buffer length in composite driver to be 4096.
Also discard control packets of size more 4KBytes in rmnet control
call back.

Change-Id: Ieb490da2e984332c66ba6764e900944cf3405e53
CRs-fixed: 280553
Signed-off-by: Vamsi Krishna <vskrishn@codeaurora.org>