msm: acpuclock-8x60: Use for_each_present_cpu for L2 voting
In the hotplug insertion path, acpuclk_set_rate() gets called
before the CPU has been brought online to the scheduler. As a
result calls to for_each_online_cpu() in this path will not
include the CPU that's being brought online

compute_l2_speed() returns an incorrect value because of this.
Replacing the use of for_each_online_cpu() in that function
with with for_each_present_cpu() resolves this.

Change-Id: I66547ede20ec0e65eeeedfb351a958080df57557
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>