msm: acpuclock-8x60: Protect L2 programming with a spinlock
It's possible for acpuclk_set_rate() to be executing concurrently
on both cores, where at least one core is executing it with
reason = SETRATE_PC.  In this case, the CPU executing with
SETRATE_PC will not acquire a mutex, since it may be executing
in an atomic context.

For SETRATE_PC, voltage and bus bandwidth setting are not
updated, so no locking is needed to protect them, but the L2
rate is updated and needs protection.  Fix this by adding a
spinlock.

Change-Id: I0c2966dfca3268cfe36a1437d9dd11d5859175b1
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>