Bluetooth: Fix problem in L2CAP ERTM receive state transition.
In cases where there are both lost packets and a slow consumer of data
(causing flow control to be triggered), the ERTM receive state machine
was not correctly transitioning back to the normal receive state once
any out-of-order data frames were reassembled.  This commit ensures
that the out-of-order queue is only checked while in the "SREJ sent"
state, and that the receive state is updated correctly once the
out-of-order queue is emptied.

Change-Id: I22a224891d137e3f39b933195c150c5840f83c2e
Signed-off-by: Mat Martineau <mathewm@codeaurora.org>