msm: smd: fix open->close->open bug
Prevent an open smd channel that is being locally closed from being
reopened by the local client until the remote has signaled a
sucessful closing of the channel.  This prevents a race condition
from occuring in the smd state machine.  Notify the client when the
channel is completely closed and ready to be reopened.

Change-Id: I29b9c48cfdbbca78905414c6afd040cb10439981
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>