camera: 8k: removed gpio reset request of front camera sensor.
The front camera sensor ov7692 does not have the reset pin connected,
the standby/power pin will do the reset. proper release of reset
for the back sensor is added. The standby pin and the reset pin for
the back sensor is pulled down rather than in the floating condition.

Change-Id: I771e2f57d5a5f518aca52873f304cdf61d750be9
Signed-off-by: Jeyaprakash Soundrapandian <jsound@codeaurora.org>