msm: kgsl: Remove race condition in GPU clk control function.
In kgsl_pwrctrl_clk(), the power_flags is set to ON after grp_clk
is enabled. The GPU interrupt could happen after the grp_clk is
enabled and before pwr_flags is set to ON. This will create the
kgsl_pwrctrl_clk re-entry situation and cause the grp_clk being
enabled twice. By setting power_flags to ON before enable the
grp_clk makes the function safe for this case.

Change-Id: Ic5d441a965fe58e039ac75af15969d789b516e02
Crs-Fixed: 279622
Signed-off-by: Wei Zou <wzou@codeaurora.org>