msm: scm: Fix incorrect cache invalidation part 2
We must invalidate the cache before reading the buffer to
determine where the response buffer begins. Otherwise we may get
stuck in a case where we think the response buffer's length is 0
(since the cache tells us the length is 0) and thus
scm_get_response_buffer() would return the same pointer as
scm_command_to_response(). This would be fine if the is_complete
field and the buf_offset field are on the same cacheline, but in
cases where they aren't we would miss invalidating the one word
we actually care about.

Instead of doing all this, just invalidate the portion of the
buffer that we really care about (the is_complete field) and then
after that bit flips invalidate the entire response buffer before
copying it to the caller. Throw in a dsb()/isb() because we don't
want the buffer getting pulled back into cache before the cache
invalidation completes.

Change-Id: I60ef54b810dba4a075d8339d24305485b8f59eff
CRs-Fixed: 281306
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>