msm: kgsl: don't call kgsl_pwrctrl_wake() from isr context
This function makes sure that the power comes on before we
try to touch the hardware via kgsl_regread and kgsl_regwrite,
which is usually a good thing.  Except that there can be races
with the isr because turning on the clocks can cause an
interrupt in some cirucumstances.

This change makes a separate set of regread/regwrite functions
that don't check that the clocks are on. These should be used
from the isr, because the clocks should already be on if we're
getting valid interrupts.

CRs-fixed: 279622
CRs-fixed: 282495
Change-Id: I4b4941a68134a4e5713ac6bb383ea8cf981d3fe0
Signed-off-by: Jeremy Gebben <jgebben@codeaurora.org>