USB: f_rmnet_sdio: Drop DTR on host suspend
Windows Hosts are not droping the DTR during suspend, but setting
the DTR as part of resume. Due to this RMNET is not getting connected.
Hence send DTR set to zero to the Modem, when the rmnet interface is
suspended.

CRs-Fixed: 280052
Signed-off-by: Velempati Chiranjeevi <cvelempa@codeaurora.org>
(cherry picked from commit c42fb7c02366cc8f6a0c3d6c2066dcdda9fbcdaf)

Change-Id: If6a2239c7d2f7586b2271e571aa797634c7000bc
Signed-off-by: Shruthi Krishna <skrish@codeaurora.org>