msm_serial: Don't configure serial interrupt as wakeup source in probe
Serial device interrupt is configured as wakeup source in probe. Due to
this, even there is no serial device user on APPS side, and same serial
instance is getting used on Modem. Serial interrupt is received by APPS
as well which is waking up the APPS. Hence configure serial interrupt as
wakeup source in startup to make sure that it is configured only when there
is user for the same.

CRs-fixed: 284567
Signed-off-by: Mayank Rana <mrana@codeaurora.org>
(cherry picked from commit 6037164912afad109bbc89d1787c57e08d7c39d1)

Change-Id: I77fc7bc3a201421d5c6b8bbd63aa7859009bd789

Signed-off-by: Sunil Joseph <sunilj@codeaurora.org>