msm: qdsp5: Initialize audio stopped flag at open
Audio stopped flag is used to track if current audio recording
session is in stopped state or not. Whenever audio recording
session is stopped or released, this flag gets set to 1.
Currently this flag is not initialized to 0 during audio open or
start. From second audio recording session onwards it is causing
failure in read() as it finds audio stopped to be already set.

Fix this by initializing audio stopped flag to 0 during audio open.

Change-Id: Ibdfbf8a7b1ef9d679b474cece7f053763dbabe87
Signed-off-by: Manish Dewangan <manish@codeaurora.org>

Conflicts:

	arch/arm/mach-msm/qdsp5/audio_voicememo.c